Rename typescriptHandler to tsComponentHandler
