+++ src/angular/index.ts
@@ -1,5 +1,5 @@
 export * from "./generateNewComponent";
 export * from "./getChanges";
 export * from "./angularTemplateHandler";
-export * from "./ts/typescriptHandler";
+export * from "./ts/tsComponentHandler";
 export * from "./isAngularCliAvailable";
