+++ src/angular/ts/index.ts
@@ -1,3 +1,3 @@
 export * from "./tsImportHandler";
 export * from "./utils";
-export * from "./typescriptHandler";
+export * from "./tsComponentHandler";
