+++ src/angular/ts/tsComponentHandler.ts
