+++ test/angular/ts/tsComponentHandler.spec.ts
